+++ src/main/java/paletteEditor/PaletteEditor.java
@@ -36,7 +36,7 @@ public class PaletteEditor
 
     private final SwatchPanel swatchPanel = new SwatchPanel();
 
-    private JComboBox<String> paletteSelector;
+    private final JComboBox<String> paletteSelector;
 
     JMenuItem menuItemPaste;
 
@@ -48,7 +48,7 @@ public class PaletteEditor
     private boolean updatingSwatches = false;
     private boolean populatingPaletteSelector = false;
 
-    private final JCheckBox desaturateButton = new javax.swing.JCheckBox("Desaturate");
+    private final JToggleButton desaturateToggleButton = new javax.swing.JCheckBox("Desaturate");
 
     private void setupMenuBar() {
         JMenuBar menuBar = new JMenuBar();
@@ -99,8 +99,31 @@ public class PaletteEditor
 
         JPanel topRowPanel = new JPanel();
         topRowPanel.setLayout(new MigLayout());
-        addPaletteSelector(topRowPanel);
-        addDesaturateButton(topRowPanel);
+
+        paletteSelector = new JComboBox<>();
+        paletteSelector.setEditable(true);
+        paletteSelector.addActionListener(this);
+        paletteSelector.addItemListener(e -> onPaletteSelected());
+        topRowPanel.add(paletteSelector);
+
+        desaturateToggleButton.addItemListener(e -> updateSongAndInstrScreens());
+        desaturateToggleButton.setToolTipText("Tip: Great palettes look OK when desaturated, too!");
+        topRowPanel.add(desaturateToggleButton);
+
+        boolean ipsScreen = EditorPreferences.getKey("ips", "0").equals("1");
+        JToggleButton ipsToggleButton = new JCheckBox("IPS");
+        ipsToggleButton.setSelected(ipsScreen);
+        ColorUtil.setIpsScreen(ipsScreen);
+        ipsToggleButton.addItemListener(e -> {
+            boolean enabled = ColorUtil.toggleIpsScreen();
+            ColorUtil.setIpsScreen(enabled);
+            EditorPreferences.putKey("ips", enabled ? "1" : "0");
+            updateSongAndInstrScreens();
+            colorPicker.repaint();
+            updateAllSwatches();
+        });
+        topRowPanel.add(ipsToggleButton);
+        topRowPanel.setToolTipText("Emulates IPS screen.");
 
         midPanel.add(topRowPanel, "grow, wrap");
         midPanel.add(colorPicker);
@@ -215,20 +238,6 @@ public class PaletteEditor
                 swatchPanel.scrollBarSwatchPair.selectForeground();
                 break;
         }
-    }
-
-    private void addPaletteSelector(JPanel panel) {
-        paletteSelector = new JComboBox<>();
-        paletteSelector.setEditable(true);
-        paletteSelector.addActionListener(this);
-        paletteSelector.addItemListener(e -> onPaletteSelected());
-        panel.add(paletteSelector);
-    }
-
-    private void addDesaturateButton(JPanel panel) {
-        desaturateButton.addItemListener(e -> updateSongAndInstrScreens());
-        desaturateButton.setToolTipText("Tip: Great palettes look OK when desaturated, too!");
-        panel.add(desaturateButton);
     }
 
     private void setRomImage(byte[] romImage) {
@@ -380,7 +389,7 @@ public class PaletteEditor
                 dstImage.setRGB(x, y, ColorUtil.colorCorrect(c));
             }
         }
-        if (desaturateButton.isSelected()) {
+        if (desaturateToggleButton.isSelected()) {
             ColorSpace colorSpace = ColorSpace.getInstance(java.awt.color.ColorSpace.CS_GRAY);
             return new java.awt.image.ColorConvertOp(colorSpace, null).filter(dstImage, dstImage);
         }